perf(ppu): sprXCounter int*→byte* + pure-SWAR slow path (+5% FPS)
Shrink sprXCounter from int[8] (32B) to byte[8] (8B) — aligns with
actual NES hardware register width. Fast-path counter==0 test now
becomes a single ulong compare.

Rewrite the slow path (when some sprite still counting down) from
an 8-iter for-loop into pure SWAR: one ulong load, one byte-wise
>0 mask, one decrement, one byte-smear, two blended shifts. No
loop, no branch, no bounds check.

Semantic proof sketch:
- dec_mask[i] = 1 iff byte v[i] > 0 (standard "is-nonzero-byte" SWAR)
- v - dec_mask has no cross-byte borrow (dec_mask[i] <= v[i])
- dec_mask * 255 ≡ (x<<8)-x byte-smears {0,1} to {0x00,0xFF}

Measured (WinForms Debug, NTSC 1x, no filter, audio 0, ny2011):
- Warm FPS   106.93 → 112.05  (+4.8%)
- Profile    105.10 → 111.32  (+5.9%)
- ppu_step_new excl  54.2% → 49.9%  (−4.3pp)
- 184/184 blargg PASS

Also adds bench_profile_1x.bat / run_perfview_1x.bat for the
"native resolution, no filter" profiling config used here.

+++ AprNes/NesCore/FDS.cs
@@ -336,7 +336,7 @@ namespace AprNes
                 for (int i = 0; i < 8; i++) { chrBankPtrs[i] = null; chrBankPtrsA[i] = null; chrBankPtrsB[i] = null; }
                 sprShiftL        = (byte*)Marshal.AllocHGlobal(sizeof(byte) * 8);
                 sprShiftH        = (byte*)Marshal.AllocHGlobal(sizeof(byte) * 8);
-                sprXCounter      = (int* )Marshal.AllocHGlobal(sizeof(int)  * 8);
+                sprXCounter      = (byte*)Marshal.AllocHGlobal(sizeof(byte) * 8);
                 sprFetchAttr     = (byte*)Marshal.AllocHGlobal(sizeof(byte) * 8);
                 sprXPos          = (byte*)Marshal.AllocHGlobal(sizeof(byte) * 8);
                 // P1_joypad_status/P2_joypad_status removed — shift register model uses static bytes
